@@ -54,10 +54,7 @@ if (pageflow.events) {
   pageflow.events.on('seed:loaded', () => boot(pageflow));
 }
 
-// `export default` does not play well with Webpack's `libraryTarget:
-// 'assign'` at the moment. See
-// https://github.com/webpack/webpack/issues/706
-module.exports = {
+export default {
   components: {
     MediaPageBackground,
     PageWithInteractiveBackground,
