@@ -20,22 +20,24 @@ var targetExternals = {
 module.exports = ['client', 'server'].map(function(target) {
   return {
     context: __dirname + '/src',
-    entry: [
-      'babel-polyfill',
-      './index.js'
-    ],
+    entry: './index.js',
 
     resolve: {
-      extensions: ['', '.js', '.jsx'],
-      root: [path.join(__dirname, 'src')]
+      extensions: ['.js', '.jsx'],
+      modules: [
+        'node_modules',
+        path.resolve(__dirname, 'src')
+      ]
     },
 
     module: {
-      loaders: [
+      rules: [
         {
           test: /\.jsx?$/,
           exclude: /node_modules/,
-          loaders: ['babel-loader'],
+          use: {
+            loader: 'babel-loader'
+          }
         }
       ],
     },
@@ -45,6 +47,7 @@ module.exports = ['client', 'server'].map(function(target) {
       filename: 'react-' + target +'.js',
 
       libraryTarget: 'assign',
+      libraryExport: 'default',
       library: ['pageflow', 'react']
     },
 
